+++ sealion/utils.py
@@ -28,7 +28,7 @@ def confusion_matrix(y_pred, y_test, plot=True):
     :return: the matrix, and show a visualization of the confusion matrix (if plot = True)
     """
 
-    y_pred, y_test = np.array(y_pred, np.int).tolist(), np.array(y_test, np.int).tolist()
+    y_pred, y_test = np.array(y_pred, int).tolist(), np.array(y_test, int).tolist()
 
     if len(y_pred) != len(y_test) : raise ValueError("y_pred and y_test must have the same size.")
 
@@ -105,7 +105,7 @@ def one_hot(indices, depth):
     :return: one-hotted features
     """
 
-    indices = np.array(indices, dtype = np.int).flatten()
+    indices = np.array(indices, dtype=int).flatten()
     zeros = np.zeros((len(indices), depth))
     df = pd.DataFrame({'col1' : zeros.tolist(), 'col2' : indices.tolist()})
     return np.apply_along_axis(_change, 1, df)
